latest

+++ src/scene/shader-lib/wgsl/chunks/gsplat/vert/gsplatSogsSH.js
@@ -6,7 +6,7 @@ uniform shN_maxs: f32;
 
 fn readSHData(source: ptr<function, SplatSource>, sh: ptr<function, array<vec3f, SH_COEFFS>>, scale: ptr<function, f32>) {
     // extract spherical harmonics palette index
-    let t: vec2<i32> = vec2<i32>(packedSample.x & 255u, packedSample.y & 255u);
+    let t: vec2<i32> = vec2<i32>(i32(packedSample.x & 255u), i32(packedSample.y & 255u));
     let n: i32 = t.x + t.y * 256;
     let u: i32 = (n % 64) * SH_COEFFS;
     let v: i32 = n / 64;
